Improved style
Only did a quick skim through the code, so I've definitely missed more
than just a few things.

1. Unless the side-effects of postfix increment operator are desired,
prefix increment should be preferred. (This is only a
micro-optimization but it is easy enough to remember and make a habit.)
2. constness should be preferred wherever possible. Functional coding
style should reduce errors.
3. Passing a deep copy of an object should be avoided. Pass by
reference or const reference instead.
4. Switch statements should compile to jump tables when case values are
close together so we shouldn't be scared of using them here. Massive
else-if ladders should be avoided.
5. We probably didn't want to use 6 separate variables for 6 mutually
exclusive conditions.
6. Reduced some function call overhead (Grid::interactMessage() in our
main).
7. Added a few small suggestions in code comments.